refactor: 관리자용 핸들러 메소드 리팩토링 및 Controller 메소드 위치 조정
