+++ server/src/main/java/server/haengdong/config/WebMvcConfig.java
@@ -40,8 +40,7 @@ public class WebMvcConfig implements WebMvcConfigurer {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(adminInterceptor())
-                .addPathPatterns("/api/**")
-                .excludePathPatterns("/api/events");
+                .addPathPatterns("/api/admin/**");
     }
 
     @Bean
