+++ server/src/main/java/server/haengdong/presentation/admin/AdminBillActionController.java
@@ -1,4 +1,4 @@
-package server.haengdong.presentation;
+package server.haengdong.presentation.admin;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -9,28 +9,30 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import server.haengdong.application.BillActionDetailService;
 import server.haengdong.application.BillActionService;
+import server.haengdong.presentation.request.BillActionDetailsUpdateRequest;
 import server.haengdong.presentation.request.BillActionUpdateRequest;
 import server.haengdong.presentation.request.BillActionsSaveRequest;
 
 @RequiredArgsConstructor
 @RestController
-public class BillActionController {
+public class AdminBillActionController {
 
     private final BillActionService billActionService;
+    private final BillActionDetailService billActionDetailService;
 
-    @PostMapping("/api/events/{eventId}/bill-actions")
+    @PostMapping("/api/admin/events/{eventId}/bill-actions")
     public ResponseEntity<Void> saveAllBillAction(
             @PathVariable("eventId") String token,
             @Valid @RequestBody BillActionsSaveRequest request
     ) {
         billActionService.saveAllBillAction(token, request.toAppRequests());
 
-        return ResponseEntity.ok()
-                .build();
+        return ResponseEntity.ok().build();
     }
 
-    @PutMapping("/api/events/{eventId}/bill-actions/{actionId}")
+    @PutMapping("/api/admin/events/{eventId}/bill-actions/{actionId}")
     public ResponseEntity<Void> updateBillAction(
             @PathVariable("eventId") String token,
             @PathVariable("actionId") Long actionId,
@@ -38,18 +40,27 @@ public class BillActionController {
     ) {
         billActionService.updateBillAction(token, actionId, request.toAppResponse());
 
-        return ResponseEntity.ok()
-                .build();
+        return ResponseEntity.ok().build();
     }
 
-    @DeleteMapping("/api/events/{eventId}/bill-actions/{actionId}")
+    @PutMapping("/api/admin/events/{eventId}/bill-actions/{actionId}/fixed")
+    public ResponseEntity<Void> updateBillActionDetails(
+            @PathVariable("eventId") String token,
+            @PathVariable("actionId") Long actionId,
+            @Valid @RequestBody BillActionDetailsUpdateRequest request
+    ) {
+        billActionDetailService.updateBillActionDetails(token, actionId, request.toAppRequest());
+
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("/api/admin/events/{eventId}/bill-actions/{actionId}")
     public ResponseEntity<Void> deleteBillAction(
             @PathVariable("eventId") String token,
             @PathVariable("actionId") Long actionId
     ) {
         billActionService.deleteBillAction(token, actionId);
 
-        return ResponseEntity.ok()
-                .build();
+        return ResponseEntity.ok().build();
     }
 }
