+++ server/src/test/java/server/haengdong/docs/BillActionDetailControllerDocsTest.java
@@ -8,12 +8,10 @@ import static org.springframework.restdocs.cookies.CookieDocumentation.cookieWit
 import static org.springframework.restdocs.cookies.CookieDocumentation.requestCookies;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
-import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.put;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
-import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
@@ -25,14 +23,11 @@ import static server.haengdong.support.fixture.Fixture.EVENT_COOKIE;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.MediaType;
 import org.springframework.restdocs.payload.JsonFieldType;
 import server.haengdong.application.BillActionDetailService;
 import server.haengdong.application.response.BillActionDetailAppResponse;
 import server.haengdong.application.response.BillActionDetailsAppResponse;
 import server.haengdong.presentation.BillActionDetailController;
-import server.haengdong.presentation.request.BillActionDetailUpdateRequest;
-import server.haengdong.presentation.request.BillActionDetailsUpdateRequest;
 
 public class BillActionDetailControllerDocsTest extends RestDocsSupport {
 
@@ -69,49 +64,6 @@ public class BillActionDetailControllerDocsTest extends RestDocsSupport {
                                 ), requestCookies(
                                         cookieWithName("eventToken").description("행사 관리자 토큰")
                                 ), responseFields(
-                                        fieldWithPath("members").type(JsonFieldType.ARRAY)
-                                                .description("전체 정산 수정 요청 목록"),
-                                        fieldWithPath("members[0].name").type(JsonFieldType.STRING)
-                                                .description("참여자 이름"),
-                                        fieldWithPath("members[0].price").type(JsonFieldType.NUMBER)
-                                                .description("참여자 정산 금액"),
-                                        fieldWithPath("members[0].isFixed").type(JsonFieldType.BOOLEAN)
-                                                .description("참여자 정산 금액 수정 여부")
-                                )
-                        )
-                );
-    }
-
-    @DisplayName("참여자별 지출 금액을 수정한다.")
-    @Test
-    void updateBillActionDetailsTest() throws Exception {
-        List<BillActionDetailUpdateRequest> billActionDetailUpdateRequests = List.of(
-                new BillActionDetailUpdateRequest("소하", 10000L, true),
-                new BillActionDetailUpdateRequest("웨디", 20000L, true)
-        );
-        BillActionDetailsUpdateRequest request = new BillActionDetailsUpdateRequest(
-                billActionDetailUpdateRequests);
-
-        String json = objectMapper.writeValueAsString(request);
-
-        mockMvc.perform(put("/api/events/{eventId}/bill-actions/{actionId}/fixed", "TOKEN", 1L)
-                        .cookie(EVENT_COOKIE)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(json))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andDo(
-                        document("updateBillActionDetailsTest",
-                                preprocessRequest(prettyPrint()),
-                                preprocessResponse(prettyPrint()),
-                                pathParameters(
-                                        parameterWithName("eventId").description("행사 ID"),
-                                        parameterWithName("actionId").description("액션 ID")
-                                ),
-                                requestCookies(
-                                        cookieWithName("eventToken").description("행사 관리자 토큰")
-                                ),
-                                requestFields(
                                         fieldWithPath("members").type(JsonFieldType.ARRAY)
                                                 .description("전체 정산 수정 요청 목록"),
                                         fieldWithPath("members[0].name").type(JsonFieldType.STRING)
