+++ server/src/test/java/server/haengdong/presentation/ControllerTestSupport.java
@@ -14,18 +14,22 @@ import server.haengdong.application.BillActionDetailService;
 import server.haengdong.application.BillActionService;
 import server.haengdong.application.EventService;
 import server.haengdong.application.MemberActionService;
+import server.haengdong.presentation.admin.AdminBillActionController;
+import server.haengdong.presentation.admin.AdminEventController;
+import server.haengdong.presentation.admin.AdminMemberActionController;
 
 @WebMvcTest(
         controllers = {
+                AdminEventController.class,
+                AdminBillActionController.class,
+                AdminMemberActionController.class,
                 EventController.class,
-                ActionController.class,
-                BillActionController.class,
                 MemberActionController.class,
                 BillActionDetailController.class
         },
         excludeFilters = {@ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = {WebMvcConfigurer.class})}
 )
-abstract class ControllerTestSupport {
+public abstract class ControllerTestSupport {
 
     @Autowired
     protected MockMvc mockMvc;
