+++ server/src/test/java/server/haengdong/presentation/EventControllerTest.java
@@ -1,11 +1,11 @@
 package server.haengdong.presentation;
 
+import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.cookie;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -15,37 +15,17 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import server.haengdong.application.request.EventAppRequest;
 import server.haengdong.application.response.EventAppResponse;
 import server.haengdong.application.response.EventDetailAppResponse;
+import server.haengdong.application.response.MemberBillReportAppResponse;
 import server.haengdong.application.response.MembersAppResponse;
 import server.haengdong.presentation.request.EventLoginRequest;
 import server.haengdong.presentation.request.EventSaveRequest;
-import server.haengdong.presentation.request.MemberNameUpdateRequest;
-import server.haengdong.presentation.request.MemberNamesUpdateRequest;
 
 
 class EventControllerTest extends ControllerTestSupport {
-
-    @DisplayName("이벤트를 생성한다.")
-    @Test
-    void saveEvent() throws Exception {
-        EventSaveRequest eventSaveRequest = new EventSaveRequest("토다리", "0987");
-        String requestBody = objectMapper.writeValueAsString(eventSaveRequest);
-        String eventId = "망쵸토큰";
-        EventAppResponse eventAppResponse = new EventAppResponse(eventId);
-        given(eventService.saveEvent(any(EventAppRequest.class))).willReturn(eventAppResponse);
-        given(authService.createToken(eventId)).willReturn("jwtToken");
-        given(authService.getTokenName()).willReturn("eventToken");
-
-        mockMvc.perform(post("/api/events")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(requestBody))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(cookie().value("eventToken", "jwtToken"))
-                .andExpect(jsonPath("$.eventId").value("망쵸토큰"));
-    }
 
     @DisplayName("토큰으로 행사를 조회한다.")
     @Test
@@ -74,22 +54,42 @@ class EventControllerTest extends ControllerTestSupport {
                 .andExpect(jsonPath("$.memberNames[1]").value("쿠키"));
     }
 
-    @DisplayName("행사 참여 인원의 이름을 수정한다.")
+    @DisplayName("참여자별 정산 현황을 조회한다.")
     @Test
-    void updateMember() throws Exception {
-        String token = "TOKEN";
-        MemberNamesUpdateRequest memberNameUpdateRequest = new MemberNamesUpdateRequest(List.of(
-                new MemberNameUpdateRequest("토다링", "토쟁이"),
-                new MemberNameUpdateRequest("감자", "고구마")
-        ));
+    void getMemberBillReports() throws Exception {
+        List<MemberBillReportAppResponse> memberBillReportAppResponses = List.of(
+                new MemberBillReportAppResponse("소하", 20_000L), new MemberBillReportAppResponse("토다리", 200_000L));
 
-        String requestBody = objectMapper.writeValueAsString(memberNameUpdateRequest);
+        given(actionService.getMemberBillReports(any())).willReturn(memberBillReportAppResponses);
 
-        mockMvc.perform(put("/api/events/{eventId}/members/nameChange", token)
+        mockMvc.perform(get("/api/events/{eventId}/reports", "망쵸토큰")
+                        .accept(MediaType.APPLICATION_JSON))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(MockMvcResultMatchers.jsonPath("$.reports[0].name").value(equalTo("소하")))
+                .andExpect(MockMvcResultMatchers.jsonPath("$.reports[0].price").value(equalTo(20_000)))
+                .andExpect(MockMvcResultMatchers.jsonPath("$.reports[1].name").value(equalTo("토다리")))
+                .andExpect(MockMvcResultMatchers.jsonPath("$.reports[1].price").value(equalTo(200_000)));
+    }
+
+    @DisplayName("이벤트를 생성한다.")
+    @Test
+    void saveEvent() throws Exception {
+        EventSaveRequest eventSaveRequest = new EventSaveRequest("토다리", "0987");
+        String requestBody = objectMapper.writeValueAsString(eventSaveRequest);
+        String eventId = "망쵸토큰";
+        EventAppResponse eventAppResponse = new EventAppResponse(eventId);
+        given(eventService.saveEvent(any(EventAppRequest.class))).willReturn(eventAppResponse);
+        given(authService.createToken(eventId)).willReturn("jwtToken");
+        given(authService.getTokenName()).willReturn("eventToken");
+
+        mockMvc.perform(post("/api/events")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(requestBody))
                 .andDo(print())
-                .andExpect(status().isOk());
+                .andExpect(status().isOk())
+                .andExpect(cookie().value("eventToken", "jwtToken"))
+                .andExpect(jsonPath("$.eventId").value("망쵸토큰"));
     }
 
     @DisplayName("행사 어드민이 로그인한다.")
