+++ server/src/test/java/server/haengdong/presentation/admin/AdminBillActionControllerTest.java
@@ -1,4 +1,4 @@
-package server.haengdong.presentation;
+package server.haengdong.presentation.admin;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doThrow;
@@ -14,11 +14,14 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import server.haengdong.exception.HaengdongErrorCode;
 import server.haengdong.exception.HaengdongException;
+import server.haengdong.presentation.ControllerTestSupport;
+import server.haengdong.presentation.request.BillActionDetailUpdateRequest;
+import server.haengdong.presentation.request.BillActionDetailsUpdateRequest;
 import server.haengdong.presentation.request.BillActionSaveRequest;
 import server.haengdong.presentation.request.BillActionUpdateRequest;
 import server.haengdong.presentation.request.BillActionsSaveRequest;
 
-class BillActionControllerTest extends ControllerTestSupport {
+class AdminBillActionControllerTest extends ControllerTestSupport {
 
     @DisplayName("지출 내역을 생성한다.")
     @Test
@@ -32,7 +35,7 @@ class BillActionControllerTest extends ControllerTestSupport {
         String requestBody = objectMapper.writeValueAsString(request);
         String eventId = "쿠키토큰";
 
-        mockMvc.perform(post("/api/events/{eventId}/bill-actions", eventId)
+        mockMvc.perform(post("/api/admin/events/{eventId}/bill-actions", eventId)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(requestBody))
                 .andDo(print())
@@ -51,7 +54,7 @@ class BillActionControllerTest extends ControllerTestSupport {
         String requestBody = objectMapper.writeValueAsString(request);
         String eventId = "소하토큰";
 
-        mockMvc.perform(post("/api/events/{eventId}/bill-actions", eventId)
+        mockMvc.perform(post("/api/admin/events/{eventId}/bill-actions", eventId)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(requestBody))
                 .andDo(print())
@@ -66,7 +69,7 @@ class BillActionControllerTest extends ControllerTestSupport {
         String requestBody = objectMapper.writeValueAsString(request);
         String eventId = "웨디토큰";
 
-        mockMvc.perform(put("/api/events/{eventId}/bill-actions/{actionId}", eventId, 1L)
+        mockMvc.perform(put("/api/admin/events/{eventId}/bill-actions/{actionId}", eventId, 1L)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(requestBody))
                 .andDo(print())
@@ -78,7 +81,7 @@ class BillActionControllerTest extends ControllerTestSupport {
     void deleteBillAction() throws Exception {
         String eventId = "토다리토큰";
 
-        mockMvc.perform(delete("/api/events/{eventId}/bill-actions/{actionId}", eventId, 1))
+        mockMvc.perform(delete("/api/admin/events/{eventId}/bill-actions/{actionId}", eventId, 1))
                 .andDo(print())
                 .andExpect(status().isOk());
     }
@@ -90,8 +93,27 @@ class BillActionControllerTest extends ControllerTestSupport {
         doThrow(new HaengdongException(HaengdongErrorCode.EVENT_NOT_FOUND))
                 .when(billActionService).deleteBillAction(any(String.class), any(Long.class));
 
-        mockMvc.perform(delete("/api/events/{eventId}/bill-actions/{actionId}", eventId, 1))
+        mockMvc.perform(delete("/api/admin/events/{eventId}/bill-actions/{actionId}", eventId, 1))
                 .andDo(print())
                 .andExpect(status().isBadRequest());
+    }
+
+    @DisplayName("참여자별 지출 금액을 수정한다.")
+    @Test
+    void updateBillActionDetailsTest() throws Exception {
+        List<BillActionDetailUpdateRequest> billActionDetailUpdateRequests = List.of(
+                new BillActionDetailUpdateRequest("소하", 10000L, true),
+                new BillActionDetailUpdateRequest("웨디", 20000L, true)
+        );
+        BillActionDetailsUpdateRequest request = new BillActionDetailsUpdateRequest(
+                billActionDetailUpdateRequests);
+
+        String json = objectMapper.writeValueAsString(request);
+
+        mockMvc.perform(put("/api/admin/events/{eventId}/bill-actions/{actionId}/fixed", "TOKEN", 1L)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(json))
+                .andDo(print())
+                .andExpect(status().isOk());
     }
 }
